Rename mister_status_server_v2.py to canonical mister_status_server.py
The previous v1 server has been removed from the repo. With v2 as the
sole implementation going forward, the _v2 suffix is no longer needed.
Update start_monitor.sh to reference the new filename.
